kvserver: move KV pretty printers to package
Epic: none
Release note: none
