+++ pkg/BUILD.bazel
@@ -273,6 +273,7 @@ ALL_TESTS = [
     "//pkg/kv/kvserver/loqrecovery/loqrecoverypb:loqrecoverypb_test",
     "//pkg/kv/kvserver/loqrecovery:loqrecovery_test",
     "//pkg/kv/kvserver/multiqueue:multiqueue_test",
+    "//pkg/kv/kvserver/print:print_test",
     "//pkg/kv/kvserver/protectedts/ptcache:ptcache_test",
     "//pkg/kv/kvserver/protectedts/ptreconcile:ptreconcile_test",
     "//pkg/kv/kvserver/protectedts/ptstorage:ptstorage_test",
@@ -1552,6 +1553,8 @@ GO_TARGETS = [
     "//pkg/kv/kvserver/loqrecovery:loqrecovery_test",
     "//pkg/kv/kvserver/multiqueue:multiqueue",
     "//pkg/kv/kvserver/multiqueue:multiqueue_test",
+    "//pkg/kv/kvserver/print:print",
+    "//pkg/kv/kvserver/print:print_test",
     "//pkg/kv/kvserver/protectedts/ptcache:ptcache",
     "//pkg/kv/kvserver/protectedts/ptcache:ptcache_test",
     "//pkg/kv/kvserver/protectedts/ptpb:ptpb",
