+++ pkg/kv/kvserver/BUILD.bazel
@@ -9,7 +9,6 @@ go_library(
         "addressing.go",
         "app_batch.go",
         "consistency_queue.go",
-        "debug_print.go",
         "doc.go",
         "flow_control_integration.go",
         "flow_control_raft_transport.go",
@@ -170,6 +169,7 @@ go_library(
         "//pkg/kv/kvserver/lockspanset",
         "//pkg/kv/kvserver/logstore",
         "//pkg/kv/kvserver/multiqueue",
+        "//pkg/kv/kvserver/print",
         "//pkg/kv/kvserver/raftentry",
         "//pkg/kv/kvserver/raftlog",
         "//pkg/kv/kvserver/rafttrace",
@@ -307,7 +307,6 @@ go_test(
         "client_test.go",
         "closed_timestamp_test.go",
         "consistency_queue_test.go",
-        "debug_print_test.go",
         "deleted_external_sstable_test.go",
         "errors_test.go",
         "flow_control_integration_test.go",
